refactor: keyButton 컴포넌트 폴더 구조 통일 및 스타일 분리
